Refactored pages structure
