+++ documentation-app/src/App.tsx
@@ -26,7 +26,7 @@ import 'maincode-ui/dist/index.css';
 import 'maincode-ui/styles/generics.css';
 import 'maincode-ui/styles/theme.css';
 import { styleScrollbars } from 'maincode-ui';
-import Page from './pages/Page';
+import PageRoutesContent from './pages/PageRoutesContent';
 
 const App: React.FC = () => {
   const themeName = useContext(ThemeContext)?.themeName;
@@ -44,7 +44,7 @@ const App: React.FC = () => {
             <Route path='/' exact={true}>
               <Redirect to='/maincode-ui/Overview' />
             </Route>
-            <Route path='/maincode-ui/' component={Page} />
+            <Route path='/maincode-ui/' component={PageRoutesContent} />
           </IonRouterOutlet>
         </IonSplitPane>
       </IonReactRouter>
