+++ src/components/copy-area/copy-area.module.css
