+++ src/components/info-area/info-area.module.css
